Add first-class int64 support for exact integer arithmetic.
Introduce int64 as a first-class numeric type alongside float64 and
complex128, enabling exact integer arithmetic and correct R7RS exact/inexact
semantics. The reader now parses integer literals as int64, arithmetic
preserves int64 when both operands are integers, and functions like
floor/ceiling/round/truncate return int64 results.

+++ gscheme/character.go
@@ -67,13 +67,13 @@ func primitiveCharDigitValue(args Pair) interface{} {
 	if arg < '0' || arg > '9' {
 		return false
 	} else {
-		return float64(arg - '0')
+		return int64(arg - '0')
 	}
 }
 
 func primitiveCharToInteger(args Pair) interface{} {
 	arg := characterConstraint(First(args))
-	return float64(arg)
+	return int64(arg)
 }
 
 func primitiveIntegerToChar(args Pair) interface{} {
